Server player count filters.
Based on #91

I ended up doing this somewhat differently. The ServerFilter model (and the DB table) were intended to be a set, so storing the actual count values in them didn't make much sense. Instead I decided to store the actual count values as a regular CVar. This also meant the code for the basic filters handling didn't need massive revamps.