(Extremely shitty) engine version redirection.
This is necessary to backport engine security fixes to old client versions. The robust engine manifest can now specify a "redirect" for a version, which the launcher will use instead of the engine version advertised by the game server.

I've already gone through and made ourselves backend infra to backport fixes to engine versions.

This change is EXTREMELY SHITTILY hacked together. The problem is basically that engine (module) versions are selected when CONTENT is downloaded (remember that the launcher basically has a 2-stage download, content then engine). I made it do redirect when the engine is subsequently downloaded and then pass up the modified engine version, but...

This introduces two problems:

* The code to cull old engine versions uses that info, which means ugly hacks are needed to make that not immediately delete the new engine version.
* Module versions are not re-resolved, which is probably gonna bite me in the ass sooner or later.

I really need to rewrite the launcher structure here, so that there's one table for "the declared versions content wants" and then have a separate database table for resolved versions. Ugh.

There's other icks. I obviously had to make the engine manifest checked even if the existing engine version is already present, but the code for caching the HTTP responses could use work. Ideally they'd be cached to disk too.